requirements.txt
fastf1>=3.3.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
scikit-learn>=1.3.0
xgboost>=2.0.0
openmeteo-requests>=1.2.0
requests-cache>=1.1.0
retry-requests>=2.0.0
newsapi-python>=0.2.7
nltk>=3.8.0
tqdm>=4.66.0
